Fixes EVENTA-SERVO-B4

+++ app/controllers/participants_controller.rb
@@ -3,14 +3,15 @@
 class ParticipantsController < ApplicationController
   def event
     event = Event.lau_ligilo(params[:event_code])
+
+    redirect_to root_url and return if event.nil?
+
     participant = event.participants.find_by(user_id: current_user.id)
 
     if participant
       participant.destroy
     else
-      event.add_participant(current_user, public: params[:publika] == 'jes')
+      event.add_participant(current_user, public: params[:publika] == "jes")
     end
-
-    redirect_to event_url(event.ligilo)
   end
 end
